requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
streamlit==1.24.1
pillow==10.0.0
pytesseract==0.3.11
torch==2.2.2
transformers==4.35.0
sentence-transformers==2.2.2
clip-anytorch==1.1.5
python-magic==0.4.27
pdf2image==1.16.3
pdfminer.six==20221105
pandas==2.1.0
